requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Bcrypt>=1.0.1
Flask-JWT-Extended>=4.6.0
Flask-Limiter>=3.5.0
Pillow>=10.1.0
gunicorn>=21.2.0
Werkzeug>=3.0.1
